chore(lint): adopt v2 lint-first selectors (N/B/SIM/UP/RUF/ERA/T20) + pyright strict on src/ (#34)
- Add select = [E/F/W/I/N/B/SIM/UP/RUF/ERA/T20] to [tool.ruff.lint]
- Add [tool.pyright] block with basic mode + strict executionEnvironment for src/
- Add [tool.ruff.lint.per-file-ignores] covering camelCase coord API (N802/N803/N806/N811/N815/N816/N818), F401 side-effect imports, B017, T201, E402
- Fix all B904 (raise-in-except missing from), B905 (zip missing strict=), B007/RUF059 (unused unpacked vars, _ prefix), B018 (property access in raises)
- Fix SIM102/SIM103/SIM105/SIM108/SIM201, RUF005/RUF015/RUF034/RUF043/RUF046 findings
- Remove dead commented-out code (ERA001 true positives); noqa false-positive ERA001 in formula/section-header comments
- Suppress RUF012 for Enum member class (CharacterGroups) via per-file-ignores
- Add E501 to global ignore (line-length enforcement not desired project-wide)

+++ pd_book_tools/ocr/block.py
@@ -838,7 +838,7 @@ class Block:
             )
             return False
 
-        character_split_index = int(round(len(word_text) * split_fraction))
+        character_split_index = round(len(word_text) * split_fraction)
         character_split_index = max(1, min(len(word_text) - 1, character_split_index))
 
         epsilon = min(1e-6, bbox_width / 10) if bbox_width > 0 else 0.0
